backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
stripe==7.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.12.1
aiosmtplib==3.0.1
email-validator==2.1.0
